requirements.txt
SpeechRecognition
PyAudio
nltk
textblob
pyAudioAnalysis
rake-nltk
yake
scipy
scikit-learn
hmmlearn
matplotlib
eyed3
pydub
imblearn
plotly
google-cloud-speech
pytorch
torch
torchvision 
torchaudio
fpdf
transformers
openai
markdown2